+++ common/constants/collectionNames.js
@@ -1,0 +1,11 @@
+const collectionNames = {
+    accessories: 'accessories',
+    assets: 'assets',
+    companies: 'companies',
+    departments: 'departments',
+    licenses: 'licenses',
+    locations: 'locations',
+    users: 'users'
+};
+
+export default collectionNames;