+++ common/constants/constats.js
@@ -1,0 +1,6 @@
+const constants = {
+    'oneSecond': 1000,
+    'tenSeconds': 10000
+};
+
+export default constants;