+++ server/src/lib/helpers/generateToken.js
@@ -1,0 +1,8 @@
+import jsonwebtoken from 'jsonwebtoken';
+
+export default (user) => {
+    const { id, isRemembered } = user;
+    return jsonwebtoken.sign({ id }, process.env.JWT_SECRET, {
+        expiresIn: isRemembered ? '30d' : '1h',
+    });
+};