+++ server/src/routes/companies.routes.js
@@ -1,0 +1,21 @@
+import { Router } from 'express';
+import {
+    getCompanies,
+    getCompany,
+    updateCompany,
+    deleteCompany,
+    createCompany,
+} from '../controllers/companies.controllers.js';
+import adminMiddleware from '../middleware/admin.middleware.js';
+
+const router = Router();
+router.route('/').get(getCompanies).post(adminMiddleware, createCompany);
+
+router
+    .route('/:id')
+    .get(getCompany)
+    .put(adminMiddleware, updateCompany)
+    .patch(adminMiddleware, updateCompany)
+    .delete(adminMiddleware, deleteCompany);
+
+export default router;