+++ server/src/routes/index.js
@@ -1,0 +1,19 @@
+import AuthRoutes from './auth.routes.js';
+import AccessoriesRoutes from './accessories.routes.js';
+import AssetsRoutes from './assets.routes.js';
+import CompaniesRoutes from './companies.routes.js';
+import DepartmentsRoutes from './departments.routes.js';
+import LicensesRoutes from './licenses.routes.js';
+import LocationsRoutes from './locations.routes.js';
+import UsersRoutes from './users.routes.js';
+
+export default {
+    AuthRoutes,
+    AccessoriesRoutes,
+    AssetsRoutes,
+    CompaniesRoutes,
+    DepartmentsRoutes,
+    LicensesRoutes,
+    LocationsRoutes,
+    UsersRoutes,
+};